Added a test that checks that gpgpwd encrypts with the same key it encrypted with
If gpgpwd can be tricked into encrypting with another key, that's technically a
security flaw (though if a user has local user access all is lost anyway), but
more importantly, if a user has several keys, this can result in the database
becoming unusable. Consider the case where a user has a machine A with two keys
and machine B with one key. If machine A decrypts with the shared key, but then
encrypts with the key only on that machine, then pushes the git repo, machine B
won't be able to read it.